修改appid为64位整型

+++ Server/KFPlugin/KFBattleClient/KFBattleClientInterface.h
@@ -14,7 +14,7 @@ namespace KFrame
         virtual bool SendMessageToBattle( uint64 roomid, uint32 msgid, ::google::protobuf::Message* message ) = 0;
 
         // ≤È—Ø’Ω≥°
-        virtual void QueryBattleRoom( uint32 playerid, uint64 roomid ) = 0;
+        virtual void QueryBattleRoom( uint64 playerid, uint64 roomid ) = 0;
     };
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////
