+++ Server/KFPlugin/KFBattleShard/KFBattleManage.h
@@ -19,8 +19,8 @@ namespace KFrame
         void Reset();
 
     public:
-        uint32 _server_id;		// 服务器id
-        uint32 _proxy_id;		// 连接的代理服务器id
+        uint64 _server_id;		// 服务器id
+        uint64 _proxy_id;		// 连接的代理服务器id
         std::string _ip;		// ip
         uint32 _port;			// port
         std::string _version;	// 版本号
@@ -39,19 +39,19 @@ namespace KFrame
         void Initialize();
 
         // 注册战场
-        void RegisterBattleServer( uint32 serverid, uint32 proxyid, const std::string& ip, uint32 port, const std::string& version );
+        void RegisterBattleServer( uint64 serverid, uint64 proxyid, const std::string& ip, uint32 port, const std::string& version );
 
         // 删除战场
-        void UnRegisterBattleServer( uint32 serverid );
+        void UnRegisterBattleServer( uint64 serverid );
 
         // 分配战场
-        void AllocBattleServer( uint32 battleserverid, const std::string& version, KFBattleServer* battleserver );
+        void AllocBattleServer( uint64 battleserverid, const std::string& version, KFBattleServer* battleserver );
 
         // 分配一个战场
-        void RemoveBattleServer( uint32 serverid, const std::string& version );
+        void RemoveBattleServer( uint64 serverid, const std::string& version );
 
         // 释放战斗服务器
-        void FreeBattleServer( uint32 serverid, const std::string& version );
+        void FreeBattleServer( uint64 serverid, const std::string& version );
 
     private:
         KFRedisDriver* _battle_redis_driver;
